refactor(node): centralize intent command resolution
